pyproject.toml
[project]
name = "conduit-router"
version = "0.0.4-alpha"
description = "ML-powered LLM routing system for cost/latency/quality optimization"
authors = [
    { name = "Evan Volgas", email = "evan@ashita.ai" }
]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["llm", "routing", "ml", "optimization", "cost-reduction"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "pydantic>=2.12.0",
    "pydantic-ai>=1.14.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "scikit-learn>=1.3.0",
    "numpy>=2.0.0",
    "redis>=5.0.0",
    "msgpack>=1.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "sqlalchemy>=2.0.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "rich>=14.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-redis>=0.41b0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "greenlet>=3.2.4",
    "psycopg2-binary>=2.9.11",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.0",
    "black>=25.0.0",
    "ruff>=0.14.0",
    "mypy>=1.18.0",
    "types-redis>=4.6.0",
]
litellm = [
    "litellm>=1.74.9",
]
arbiter = [
    "arbiter-ai>=0.1.0",
]
embeddings = [
    "sentence-transformers>=2.2.0",
    "openai>=1.0.0",
]

[project.scripts]
conduit = "conduit.cli.main:cli"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["conduit*", "conduit_litellm*"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312", "py313"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "E402",  # module level import not at top of file (load_dotenv needs to run first)
    "F401",  # imported but unused (re-exports in __init__.py)
    "F823",  # local variable referenced before assignment (PCA conditional import)
    "N803",  # argument name should be lowercase (matrices use math notation: A, L, etc.)
    "N806",  # variable in function should be lowercase (matrices use math notation)
    "SIM105",  # use contextlib.suppress (explicit try-except-pass is clearer)
    "SIM108",  # use ternary operator (if-else blocks are clearer for complex logic)
    "B905",  # zip without strict parameter (Python 3.10+ only, maintaining 3.10 compat)
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

# Ignore missing stubs for third-party libraries without type hints
[[tool.mypy.overrides]]
module = ["fastapi.*", "uvicorn.*", "sentence_transformers.*", "sklearn.*", "openai.*"]
ignore_missing_imports = true
disallow_untyped_decorators = false

[tool.pytest.ini_options]
minversion = "9.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"
# For coverage: uv run pytest --cov=conduit --cov-report=term-missing

[dependency-groups]
dev = [
    "black>=25.11.0",
    "mypy>=1.18.2",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.6",
    "types-pyyaml>=6.0.12.20250915",
    "matplotlib>=3.10.0",
    "pillow>=12.0.0",
    "pyparsing>=3.2.0",
]
